requirements.txt
fastapi==0.68.0
uvicorn==0.12.2
PyYAML==6.0
tweepy==4.13.0
aiohttp==3.8.4
facebook-sdk==3.1.0